% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{chunker}
\alias{chunker}
\title{Build an chunker}
\usage{
chunker(features_chunked, classifier = randomForest::randomForest, ...)
}
\arguments{
\item{features_chunked}{Features of chunked text}

\item{classifier}{Defaults to randomForest}

\item{...}{Extra parameters to pass to randomForest}
}
\description{
Build an chunker based on random forest to chunk text with
    institutional grammar
}
\examples{
chunked = data.frame(
  document = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  text = c("Power plants", "must not", "ever", "pollute", "the air",
            "and also",
            "sewage plants", "must not", "ever", "pollute", "the water"),
  component = c("attribute", "deontic", NA, "aim", "object",
                NA,
                "attribute", "deontic", NA, "aim", "object"),
  statement_ID = c(1, 1, 1, 1, 1,
                   2,
                   3, 3, 3, 3, 3)
)
unchunked = data.frame(
  document = 2,
  text = "Chemical plants must not ever pollute the soil"
)
tags = tag(chunked, unchunked)
my_features = features(tags)
chunker(my_features$chunked)

}
