% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{features}
\alias{features}
\title{Create features}
\usage{
features(tags, number_of_words = 10, part_of_speech_window = 20,
  word_window = 5)
}
\arguments{
\item{tags}{Tagged data, the result of tag.}

\item{number_of_words}{The number of distinct words to use for chunking}

\item{part_of_speech_window}{The number of parts of speech before and after
each word to use for chunking}

\item{word_window}{The number of words before and after each word to use for
chunking}
}
\description{
Create features for chunking from tagged data
}
\examples{
chunked = data.frame(
  document = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  text = c("Power plants", "must not", "ever", "pollute", "the air",
            "and also",
            "sewage plants", "must not", "ever", "pollute", "the water"),
  component = c("attribute", "deontic", NA, "aim", "object",
                NA,
                "attribute", "deontic", NA, "aim", "object"),
  statement_ID = c(1, 1, 1, 1, 1,
                   2,
                   3, 3, 3, 3, 3)
)
unchunked = data.frame(
  document = 2,
  text = "Chemical plants must not ever pollute the soil"
)
tags = tag(chunked, unchunked)
features(tags)

}
