% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{parts_of_speech}
\alias{parts_of_speech}
\title{Create parts of speech}
\usage{
parts_of_speech(chunked, unchunked)
}
\arguments{
\item{chunked}{A data frame with a text variable.}

\item{unchunked}{A data frame with a text variable.}
}
\description{
Chunked and unchunked text must be included together.
}
\examples{
chunked = data.frame(
  document = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  text = c("Power plants", "must not", "ever", "pollute", "the air",
            "and also",
            "sewage plants", "must not", "ever", "pollute", "the water"),
  component = c("attribute", "deontic", NA, "aim", "object",
                NA,
                "attribute", "deontic", NA, "aim", "object"),
  statement_ID = c(1, 1, 1, 1, 1,
                   2,
                   3, 3, 3, 3, 3)
)
unchunked = data.frame(
  document = 2,
  text = "Chemical plants must not ever pollute the soil"
)
parts_of_speech(chunked, unchunked)

}
