% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{training_and_testing}
\alias{training_and_testing}
\title{Split data into training and testing data}
\usage{
training_and_testing(data, fraction = 0.5)
}
\arguments{
\item{data}{A data frame}

\item{fraction}{The percentage of the data to put into the training data}
}
\description{
Will return a list with a training and a testing data
}
\examples{
chunked = data.frame(
  document = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  text = c("Power plants", "must not", "ever", "pollute", "the air",
            "and also",
            "sewage plants", "must not", "ever", "pollute", "the water"),
  component = c("attribute", "deontic", NA, "aim", "object",
                NA,
                "attribute", "deontic", NA, "aim", "object"),
  statement_ID = c(1, 1, 1, 1, 1,
                   2,
                   3, 3, 3, 3, 3)
)
unchunked = data.frame(
  document = 2,
  text = "Chemical plants must not ever pollute the soil"
)
my_features = features(chunked, unchunked, number_of_words = 3, window = 3)
training_and_testing(my_features$chunked, fraction = 0.5)

}
