% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{balance}
\alias{balance}
\title{Balance data}
\usage{
balance(data, minimum_number = 100)
}
\arguments{
\item{data}{A data frame with a tag column}

\item{minimum_number}{A minimum number of observations per tag}
}
\description{
Will discard tags with below the minimum number of observations, and sample
down the rest of the tags to have the same number of observations.
}
\examples{
chunked = data.frame(
  document = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  text = c("Power plants", "must not", "ever", "pollute", "the air",
            "and also",
            "sewage plants", "must not", "ever", "pollute", "the water"),
  component = c("attribute", "deontic", NA, "aim", "object",
                NA,
                "attribute", "deontic", NA, "aim", "object"),
  statement_ID = c(1, 1, 1, 1, 1,
                   2,
                   3, 3, 3, 3, 3)
)
unchunked = data.frame(
  document = 2,
  text = "Chemical plants must not ever pollute the soil"
)
tags = tag(chunked, unchunked)
my_features = features(tags)
balance(my_features$chunked, minimum_number = 0)

}
