% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{chunk}
\alias{chunk}
\title{Chunk text}
\usage{
chunk(chunker, features_unchunked)
}
\arguments{
\item{chunker}{A chunker}

\item{features_unchunked}{Unchunked text features}
}
\description{
Chunk text with a chunker and unchunked features.
}
\examples{
chunked = data.frame(
  source = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  text = c("Power plants", "must not", "ever", "pollute", "the air",
            "and also",
            "sewage plants", "must not", "ever", "pollute", "the water"),
  component = c("attribute", "deontic", NA, "aim", "object",
                NA,
                "attribute", "deontic", NA, "aim", "object"),
  statement_ID = c(1, 1, 1, 1, 1,
                   2,
                   3, 3, 3, 3, 3)
)
unchunked = data.frame(
  source = 2,
  text = "Chemical plants must not ever pollute the soil"
)
features = features(chunked, unchunked, number_of_words = 3, window = 3)
chunker = chunker(features$chunked)
chunk(chunker, features$unchunked)

}
