% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{features}
\alias{features}
\title{Create features for chunking}
\usage{
features(chunked, unchunked, number_of_words = 50, window = 10)
}
\arguments{
\item{chunked}{A data frame of chunked text. Include four columns: document (a
document text ID), component (one of ABDICO, if applicable), text (text
of the fragment), and a statement_ID for each statement and segment
of text between statements.}

\item{unchunked}{A dataframe of unchunked text. Include two columns:
document (a document text ID) and text (the contexts of the text).}

\item{number_of_words}{The number of distinct words to use for chunking}

\item{window}{The number of words before and after each word to use for
chunking}
}
\description{
Chunked and unchunked text must be included together. Will return a list
with features for both.
}
\examples{
chunked = data.frame(
  document = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  text = c("Power plants", "must not", "ever", "pollute", "the air",
            "and also",
            "sewage plants", "must not", "ever", "pollute", "the water"),
  component = c("attribute", "deontic", NA, "aim", "object",
                NA,
                "attribute", "deontic", NA, "aim", "object"),
  statement_ID = c(1, 1, 1, 1, 1,
                   2,
                   3, 3, 3, 3, 3)
)
unchunked = data.frame(
  document = 2,
  text = "Chemical plants must not ever pollute the soil"
)
features(chunked, unchunked, number_of_words = 3, window = 3)

}
