% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{validate}
\alias{validate}
\title{Validate a chunker}
\usage{
validate(chunker, testing)
}
\arguments{
\item{chunker}{a chunker}

\item{testing}{testing data features}
}
\description{
Will return a two way table to assess tagging accuracy. Actual tags will be
rows, and predicted tags will be columns.
}
\examples{
chunked = data.frame(
  document = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  text = c("Power plants", "must not", "ever", "pollute", "the air",
            "and also",
            "sewage plants", "must not", "ever", "pollute", "the water"),
  component = c("attribute", "deontic", NA, "aim", "object",
                NA,
                "attribute", "deontic", NA, "aim", "object"),
  statement_ID = c(1, 1, 1, 1, 1,
                   2,
                   3, 3, 3, 3, 3)
)
unchunked = data.frame(
  document = 2,
  text = "Chemical plants must not ever pollute the soil"
)
tags = tag(chunked, unchunked)
my_features = features(tags)
my_training_and_testing = training_and_testing(my_features$chunked)
my_chunker = chunker(my_training_and_testing$training)
validate(my_chunker, my_training_and_testing$testing)

}
